requirements.txt
discord.py
aiohttp==3.7.4
asyncio==3.4.3
urllib3==1.26.5
lxml==4.6.5
SQLAlchemy==1.4.27
pyyaml==6.0
googletrans==3.1.0a0
aioschedule==0.5.2
tabulate==0.8.9
gTTS==2.2.3
psycopg2==2.9.3